Fix for avoiding division-by-zero cases on iOS/Metal.
On the original code, the XYZ-xyY functions are taking max between 1e-10
to avoid division-by-zero, but it doesn't take effect on iOS/Metal
because of rounding errors, then they cause division-by-zero when nearly
black colors are given.

I pulled them up to 1e-4 because 1e-5 was not enough as far as I tried.