DoF quality improvements when Bloom is enabled
This comes at the price of an additional (quite cheap) pass but the
visual quality gain is quite huge, it's well worth the small performance
loss. On the other hand, when Chromatic Aberration are enabled,
performances are actually increased because DoF only needs to be
resolved once (instead of once per CA sample).