Example usage of XRSettings.eyeTextureDesc
There is no stereo instancing in 2017.2 on PC, so you probably can ignore the rtArray bits for now (unless you want Hololens supported right now).

The weird SetRenderTarget and Blit stuff is because Blit only copies one slice at a time.  We're going to add the functionality for multiple slices to Blit for 2017.3.