First pass at AutoExposure XR update
This properly supports multi-pass XR.
For single-pass double-wide, there is the question of a unified histogram buffer across both eyes.  It's not entirely clear to me that we really need individual histogram buffers per eye in single-pass.

Additionally, the LogHistogram workflow has to be updated to at least support Tex2DArray.

Also, I just realized we might need to add Tex2DMS support.  I wonder at what point in the pipeline we want to resolve the texture, and hand it off to post-processing to operate on.