Support RenderViewportScale
Traditional screen-space RenderViewportScale support is done through _MainTex_ST.  However, we don't use Graphics.Blit or CommandBuffer.Blit for most blits, so we can't rely on that infrastructure.

Instead, we can wrap our usage of unity_StereoScaleOffset/TransformStereoScreenSpaceTex with an RVS scaler.  This works pretty well, and makes the RenderViewportScale management very explicit and obvious.

There might be an issue with WindowsMR, as it doesn't guarantee the RenderViewportScale value will actually be used (yeah...), but we'll cross that bridge when we get there.