New PostProcessRenderContext.GetScreenSpaceTemporaryRT
We need XRSettings.eyeTextureDesc to create proper single-pass temporary RTs.  But they aren't available until 2017.2.  So if there is the need to create a temporary screen space RT, there is a new function that handles the allocation.

It either uses the old explicit-argument-driven GetTemporaryRT for pre-2017.2 (monoscopic + multi-pass) or RenderTextureDescriptor 2017.2+.  There is also a 'scaler' that is applied to the texture dimensions (if needed).  I might make this more explicit in the future with explicit arguments changing the size.