Fix inconsistent resource stripping
OnPostprocessBuild is not called if a build is cancelled.
This caused inconsistencies because the modification to the resource list
are not reverted in that case.
Now we have two separate PostProcessResources assets:
an unstripped complete list of resources that is only used in Editor
and a separate asset that is generated from the complete list during stripping.
The runtime only references the stripped asset.
Also the stripping configuration is moved into a separate asset.